app_pages_yellow/requirements.txt
asgiref==3.2.10
autopep8==1.5.7
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.2
colorama==0.4.4
coverage==6.3.2
cprint==1.1
cryptography==3.2
cycler==0.10.0
defusedxml==0.7.0rc1
distlib==0.3.4
Django==4.2.9
django-braces==1.14.0
django-cors-headers==3.4.0
django-crontab==0.7.1
django-filter==2.4.0
django-mathfilters==1.0.0
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.6.0
filelock==3.6.0
flake8==4.0.1
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==4.12.0
kiwisolver==1.2.0
matplotlib==3.3.1
mccabe==0.6.1
mypy-extensions==0.4.3
netaddr==0.8.0
numpy==1.21.6
oauthlib==3.1.0
pandas==1.3.5
pathspec==0.9.0
Pillow==8.0.0
platformdirs==2.5.1
protobuf==3.14.0
psycopg==3.0.10
psycopg2==2.9.3
psycopg2-binary==2.8.6
pybase64==1.1.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.9.9
pycryptodomex==3.9.9
pyfcm==1.4.7
pyflakes==2.4.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
reportlab==3.5.65
requests==2.24.0
requests-oauthlib==1.3.0
schedule==0.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.3
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==3.10.0.0
tzdata==2022.2
urllib3==1.25.10
virtualenv==20.13.3
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
